fix(database): corrige falta de precisao nos valores bancarios

+++ api/src/database/migrations/20221116204118-create-account.js
@@ -11,8 +11,8 @@ module.exports = {
       },
       balance: {
         allowNull: false,
-        type: Sequelize.INTEGER,
-        defaultValue: 100, 
+        type: Sequelize.DECIMAL(15,2),
+        defaultValue: 100.00, 
       }
     });
   },
